feat: Move logout button to profile dropdown
- I moved the logout button and its corresponding `handleLogout` function from `TopHeader.jsx` to `ProfileDropdown.jsx`.
- I styled the logout button in `ProfileDropdown.jsx`.
